Move `springNoConfigApplicationContext` constant from `utbot-testing` to `utbot-spring-test`
